+++ src/Components/CreateImageWizard/steps/Oscap/index.tsx
@@ -10,7 +10,8 @@ import {
 } from '@patternfly/react-core';
 import { ExternalLinkAltIcon } from '@patternfly/react-icons';
 
-import { Oscap, removeBetaFromRelease } from './Oscap';
+import { Oscap } from './Oscap';
+import { removeBetaFromRelease } from './removeBetaFromRelease';
 
 import {
   COMPLIANCE_AND_VULN_SCANNING_URL,
