+++ src/test/Components/CreateImageWizard/steps/Kernel/Kernel.test.tsx
@@ -134,9 +134,9 @@ const removeKernelArg = async (kernelArg: string) => {
 
 const selectProfile = async () => {
   const user = userEvent.setup();
-  const selectProfileDropdown = await screen.findByRole('textbox', {
-    name: /select a profile/i,
-  });
+  const selectProfileDropdown = await screen.findByPlaceholderText(
+    /select a profile/i
+  );
   await waitFor(() => user.click(selectProfileDropdown));
 
   const cis1Profile = await screen.findByText(/Kernel append only profile/i);
