+++ src/test/Components/CreateImageWizard/steps/Oscap/Oscap.test.tsx
@@ -59,9 +59,9 @@ const selectWslTarget = async () => {
 
 const selectProfile = async () => {
   const user = userEvent.setup();
-  const selectProfileDropdown = await screen.findByRole('textbox', {
-    name: /select a profile/i,
-  });
+  const selectProfileDropdown = await screen.findByPlaceholderText(
+    /select a profile/i
+  );
   await waitFor(() => user.click(selectProfileDropdown));
 
   const cis1Profile = await screen.findByText(
@@ -72,9 +72,9 @@ const selectProfile = async () => {
 
 const selectDifferentProfile = async () => {
   const user = userEvent.setup();
-  const selectProfileDropdown = await screen.findByRole('textbox', {
-    name: /select a profile/i,
-  });
+  const selectProfileDropdown = await screen.findByPlaceholderText(
+    /select a profile/i
+  );
   await waitFor(() => user.click(selectProfileDropdown));
 
   const cis2Profile = await screen.findByText(
@@ -85,9 +85,9 @@ const selectDifferentProfile = async () => {
 
 const selectNone = async () => {
   const user = userEvent.setup();
-  const selectProfileDropdown = await screen.findByRole('textbox', {
-    name: /select a profile/i,
-  });
+  const selectProfileDropdown = await screen.findByPlaceholderText(
+    /select a profile/i
+  );
   await waitFor(() => user.click(selectProfileDropdown));
 
   await waitFor(async () => user.click(await screen.findByText(/none/i)));
@@ -188,9 +188,9 @@ describe('Step OpenSCAP', () => {
     await screen.findByText(
       /OpenSCAP profiles are not compatible with WSL images/i
     );
-    expect(
-      await screen.findByRole('textbox', { name: /select a profile/i })
-    ).toBeDisabled();
+    expect(await screen.findByTestId('profileSelect')).toHaveClass(
+      'pf-m-disabled'
+    );
   });
 
   test('alert displayed and OpenSCAP dropdown enabled when targets include WSL', async () => {
@@ -203,9 +203,7 @@ describe('Step OpenSCAP', () => {
       /OpenSCAP profiles are not compatible with WSL images/i
     );
     await waitFor(() => {
-      expect(
-        screen.getByRole('textbox', { name: /select a profile/i })
-      ).toBeEnabled();
+      expect(screen.getByPlaceholderText(/select a profile/i)).toBeEnabled();
     });
   });
